+++ deps/deno_module_resolver.ts
@@ -1,0 +1,1 @@
+export * from "jsr:@miyauci/deno-module-resolver@1.0.0-beta.2";
